Javascript restructure
