+++ resources/assets/js/back.app.js
@@ -21,7 +21,7 @@ new Vue({
         media(resolve) {
             require.ensure([], require => {
                 resolve(require('blender-media').default);
-            }, 'front.blender-media');
+            }, 'back.blender-media');
         },
     },
 });
@@ -29,12 +29,11 @@ new Vue({
 // Heavy components coming up
 
 if ($('[data-chart]').size()) {
-    require.ensure([], () => { require('./modules/chart').default(); }, 'back.chart');
+    require.ensure([], () => { require('blender/chart').default(); }, 'back.chart');
 }
 
 // Uncomment if redactor files are present
 
 //if ($('[data-editor]').size()) {
-//    require.ensure([], () => { require('./modules/editor').default(); }, 'back.editor');
+//    require.ensure([], () => { require('blender/modules/editor').default(); }, 'back.editor');
 //}
-
