+++ resources/assets/js/back.head.js
@@ -1,9 +1,9 @@
-import webfont from 'webfontloader';
 import 'browsernizr/test/css/animations';
 import 'browsernizr/test/css/transforms3d';
 import 'browsernizr/test/svg';
 import 'browsernizr/test/touchevents';
 import 'browsernizr';
+import webfont from 'webfontloader';
 
 webfont.load({
     custom: {
