+++ resources/assets/js/blender/editor.js
