+++ webpack.config.js
@@ -1,14 +1,15 @@
+const path = require('path');
 const webpack = require('webpack');
 
 const config = require('blender-gulp/config/webpack');
 
 config.entry = {
     'back.vendor': ['jquery'],
-    'back.head': './js/back/head.js',
-    'back.app': './js/back/app.js',
+    'back.head': './js/back.head.js',
+    'back.app': './js/back.app.js',
     'back.style': './sass/back/back.scss',
-    'front.head': './js/front/head.js',
-    'front.app': './js/front/app.js',
+    'front.head': './js/front.head.js',
+    'front.app': './js/front.app.js',
     'front.style': './sass/front/front.scss',
 };
 
@@ -22,5 +23,7 @@ config.plugins.push(new webpack.ProvidePlugin({
     $: 'jquery',
     jQuery: 'jquery',
 }));
+
+config.resolve.root = path.resolve('resources/assets/js');
 
 module.exports = config;
